First commit of bash_scripts repo
lcl is a command to execute another command 'locally' -- that is, on the
current physical machine. It is far from perfect, but should work in
most cases (one hop of ssh)

log is a command to log another command, both to the console and to
files (one for combined stdout and stderr, and two for each of them separately). It depends on lcl.

test_alert is a demonstration of how to make alert sounds in this
environment

test_log is a demonstration and test of log. $ log test_log to see more